test(reconstitution): inline the cold-clone harness into the test file (#2272)
Fold the fixture builder + flow driver from reconstitution-harness.ts into
cold-clone-reconstitution.test.ts and delete the standalone harness module
(+ its test). The harness was a plain `src/` .ts, so it was compiled into
dist/ and shipped as dead test-only code in @deftai/directive-core; keeping it
inside the excluded *.test.ts file ships nothing extra. The #2272 xBRIEF lists
the harness as optional ("if shared setup is non-trivial"); the acceptance
coverage (fixture self-checks + a1-a4) is unchanged. Addresses the SLizard P1.

Refs #2272, #2203
